refactored location
