+++ src/test/java/com/codecool/shop/dao/implementation/SupplierDaoMemTest.java
@@ -1,5 +1,6 @@
-package com.codecool.shop.dao;
+package com.codecool.shop.dao.implementation;
 
+import com.codecool.shop.dao.SupplierDao;
 import com.codecool.shop.dao.implementation.SupplierDaoMem;
 import com.codecool.shop.model.Supplier;
 import org.junit.jupiter.api.BeforeEach;
